requirements.txt
Flask
python-dotenv
requests
transformers
torch
flask_cors
google-generativeai